refector: Type001~N까지 여러 스타일을 지원할 수 있는 형태로 변경
