+++ online_invitaion/db.json
@@ -1,6 +1,108 @@
 {
-  "wedding": {
-    "id": 0,
+  "invitation": [
+    {
+    "id": "0001",
+    "type": "Type001",
+    "date": "2024-01-27T13:00:00",
+    "location": {
+      "lat": 37.28163212324522,
+      "lng": 127.0303329958705,
+      "name": "노블레스 웨딩컨벤션 7층 컨벤션홀",
+      "address": "경기도 수원시 팔달구 우만동 팔달문로 128\n수원 노블레스 웨딩컨벤션",
+      "link": "https://map.kakao.com/?map_type=TYPE_MAP&target=car&rt=,,506715,1050669&rt1=&rt2=%EC%88%98%EC%9B%90%EB%85%B8%EB%B8%94%EB%A0%88%EC%8A%A4%EC%9B%A8%EB%94%A9%EC%BB%A8%EB%B2%A4%EC%85%98&rtIds=,9846217",
+      "waytocome": {
+        "metro": [
+          "셔틀버스 운행 → 분당선 수원시청역 5번출구 수시운행"
+        ],
+        "bus": [
+          "수원역(4번출구) → 동수원병원 하차10, 11-1, 37, 720-2, 83-1 (약 20분)",
+          "수원종합버스터미널 → 수병원 하차300, 300-1, 82-1, 80, 88 (약 20분)",
+          "서수원시외버스터미널 → 동수원병원 하차11-1, 37, 61, 62-1 (약 30분)",
+          "망포역 (4번출구) → 동수원병원 하차62-1, 61 (약 30분)",
+          "용인시(용인대입구삼거리) → 동수원병원 하차10, 10-5, 66 (약 1시간)",
+          "범계역(범계사거리), 오산역(오산터미널맞은편) → 수병원 하차300 (약 50분)",
+          "잠실역(4번출구) → 월드컵경기장 하차1007-1 (약 1시간 20분)",
+          "사당역(4번출구) → 월드컵경기장 하차7000, 7001 (약 1시간)",
+          "강남역(7번출구) → 월드컵경기장 하차3002, 3007, 3008 (약 1시간)"
+        ]
+      }
+    },
+    "groom": {
+      "name": "류우선",
+      "account": {
+        "bankName": "신한",
+        "accountNumber": "110-356-123433",
+        "kakaopayLink": "https://qr.kakaopay.com/FPV2n0zug"
+      },
+      "phoneNumber": "01045532323",
+      "parents": [
+        {
+          "name": "류영만",
+          "account": {
+            "bankName": "농협",
+            "accountNumber": "233331-51-061400"
+          },
+          "phoneNumber": "01053242422"
+        },
+        {
+          "name": "봉미선",
+          "account": {
+            "bankName": "농협",
+            "accountNumber": "352-1252-4559-90"
+          },
+          "phoneNumber": "01054912252"
+        }
+      ]
+    },
+    "bride": {
+      "name": "이지혜",
+      "account": {
+        "bankName": "농협",
+        "accountNumber": "352-0952-4669-11"
+      },
+      "phoneNumber": "01022345124",
+      "parents": [
+        {
+          "name": "이훈이",
+          "account": {
+            "bankName": "카카오뱅크",
+            "accountNumber": "3333-1112-11932"
+          },
+          "phoneNumber": "01024333423"
+        },
+        {
+          "name": "정유리",
+          "account": {
+            "bankName": "국민",
+            "accountNumber": "222-19-8842-472"
+          },
+          "phoneNumber": "01099928884"
+        }
+      ]
+    },
+    "message": {
+      "intro": "가장 진실된 것으로\n가장 당신을 위한 생각으로\n나의 마음을 가득 채워가고 싶다\n참 작은 마음이지만\n당신을 위한 것 중 가장 작은 것일지라도\n당신이 허락해준다면\n나 내 온 마음 그것이라 하겠다.\n\n이경선, <마음>",
+      "invitation": "소중한 분들을 초대합니다\n살랑이는 바람결에\n사랑이 묻어나는 계절입니다.\n여기 곱고 예쁜 두 사람이 사랑을 맺어\n인생의 반려자가 되려 합니다.\n새 인생을 시작하는 이 자리에 오셔서\n축복해 주시면 감사하겠습니다."
+    },
+    "galleryImages": [
+      "/assets/images/wedding_01",
+      "/assets/images/wedding_02",
+      "/assets/images/wedding_03",
+      "/assets/images/wedding_04",
+      "/assets/images/wedding_05",
+      "/assets/images/wedding_06",
+      "/assets/images/wedding_07",
+      "/assets/images/wedding_08",
+      "/assets/images/wedding_09",
+      "/assets/images/wedding_10",
+      "/assets/images/wedding_11",
+      "/assets/images/wedding_12"
+    ],
+    "attendCount": 2325644
+  },
+  {
+    "id": "0002",
+    "type": "Type002",
     "date": "2024-01-27T13:00:00",
     "location": {
       "lat": 37.28163212324522,
@@ -98,4 +200,5 @@
     ],
     "attendCount": 2325644
   }
+  ]
 }
