+++ online_invitaion/public/index.html
@@ -29,7 +29,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <div id="root-portal"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
