+++ online_invitaion/src/components/shared/commons/FullScreenMessage.module.scss
@@ -1,4 +1,10 @@
 .container {
+    --red: #ea7664;
+    --black: #544f4f;
+    --brown: #89757a;
+    --beige: #f6f5f5;
+    
+
     position: fixed;
     top: 0;
     right: 0;
