+++ online_invitaion/src/components/shared/dimmeds/Dimmed.module.scss
@@ -6,7 +6,7 @@
     left: 0;
     right: 0;
     bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5);
+    background-color: rgba(0, 0, 0, 0.9);
     z-index: $dimmed-zindex;
 
 }
