+++ online_invitaion/src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
-import { ModalContext } from './contexts/ModalContext'
 import reportWebVitals from './reportWebVitals'
 
 import './scss/global.scss'
@@ -9,9 +8,7 @@ import './scss/global.scss'
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
-    <ModalContext>
-      <App />
-    </ModalContext>
+    <App />
   </React.StrictMode>,
 )
