+++ online_invitaion/src/pages/Type001/Type001.module.scss
@@ -1,0 +1,8 @@
+.container{
+    --red: #ea7664;
+    --black: #544f4f;
+    --brown: #89757a;
+    --beige: #f6f5f5;
+    
+    font-family: 'NanumGeumEunBoHwa';
+}
