+++ online_invitaion/src/scss/global.scss
@@ -3,14 +3,3 @@
 @import url("./reset.scss"); // 리셋 css
 @import url("./font.scss"); // 폰트불러오기
 
-:root{
-    --red: #ea7664;
-    --black: #544f4f;
-    --brown: #89757a;
-    --beige: #f6f5f5;
-}
-
-
-body{
-    font-family: 'NanumGeumEunBoHwa';
-}
